Extraindo arquivo de rotas
